refactor: consolidate duplicated utilities and doc-indexing logic
Extract downloadBlob, heading-slug, email validation, Shiki helpers,
and doc-indexing/frontmatter parsing into shared modules. Unify usage
across components, scripts, and the MCP. No breaking changes.
