+++ package.json
@@ -11,8 +11,8 @@
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
     "generate:openapi": "npx tsx scripts/generate-openapi.ts",
-    "docs:index": "node scripts/generate-docs-index.mjs",
-    "prebuild": "node scripts/generate-docs-index.mjs"
+    "docs:index": "npx tsx scripts/generate-docs-index.ts",
+    "prebuild": "npx tsx scripts/generate-docs-index.ts"
   },
   "dependencies": {
     "@supabase/supabase-js": "^2.97.0",
