+++ src/app/api/privacy/delete/route.ts
@@ -1,9 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { supabase, isSupabaseConfigured } from '@/lib/supabase';
-
-function isValidEmail(value: string): boolean {
-  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(value);
-}
+import { isValidEmail } from '@/utils/email-validation';
 
 export async function POST(req: NextRequest) {
   const body = await req.json().catch(() => null);
