+++ src/components/docs/CodeBlock.tsx
@@ -4,6 +4,7 @@ import { useState, useEffect, useRef } from "react";
 import { Copy, Check, Code2 } from "lucide-react";
 import { cn } from "@/lib/cn";
 import { useTheme } from "@/components/providers/ThemeProvider";
+import { highlightCache, escapeHtml } from "@/utils/shiki-helpers";
 
 interface CodeBlockProps {
   code?: string;
@@ -20,18 +21,6 @@ const LANGUAGE_ALIASES: Record<string, string> = {
   conf: "ini",
   config: "ini",
 };
-
-// Global cache — persists across renders and component instances
-const highlightCache = new Map<string, string>();
-
-function escapeHtml(text: string) {
-  return text
-    .replaceAll("&", "&amp;")
-    .replaceAll("<", "&lt;")
-    .replaceAll(">", "&gt;")
-    .replaceAll('"', "&quot;")
-    .replaceAll("'", "&#039;");
-}
 
 export function CodeBlock({
   code: codeProp,
