+++ src/components/docs/ExportJSON.tsx
@@ -3,6 +3,8 @@
 import { useCallback, useMemo } from "react";
 import { FileJson } from "lucide-react";
 
+import { downloadBlob } from "@/utils/downloadBlob";
+
 interface ExportJSONProps {
   /**
    * The documentation slug of the current page.
@@ -84,18 +86,6 @@ export interface ExportedCodeBlockJSON {
    * Raw code contents from the <code> element.
    */
   code: string;
-}
-
-function downloadBlob(filename: string, content: string, mimeType: string) {
-  const blob = new Blob([content], { type: mimeType });
-  const url = URL.createObjectURL(blob);
-  const link = document.createElement("a");
-  link.href = url;
-  link.download = filename;
-  document.body.appendChild(link);
-  link.click();
-  link.remove();
-  URL.revokeObjectURL(url);
 }
 
 function getSectionNodes(allNodes: Element[], startIndex: number): Element[] {
