+++ src/utils/email-validation.ts
@@ -1,0 +1,8 @@
+// Shared utility: email validation regex and helper function.
+
+// Compiled email regex used consistently across client and server.
+export const EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+
+export function isValidEmail(value: string): boolean {
+  return EMAIL_REGEX.test(value);
+}
