+++ src/utils/heading-slug.ts
@@ -1,0 +1,9 @@
+// Shared utility: generate a URL-safe slug from heading text.
+
+export function slugifyHeading(text: string): string {
+  return text
+    .toLowerCase()
+    .replace(/[^a-z0-9\s-]/g, "")
+    .trim()
+    .replace(/\s+/g, "-");
+}
